Added execute method as the preferred way to execute actions. It forwards the head branch, and returns the executed action using its actual type
From TicTacToe testing project:
Right now, actions, not states, are transferred, so both [engines] have different state IDs and are strictly speaking not synchronized. To fix this, a more intricate protocol is needed that is aware of different branches, so that it can detect where deserialized states belong. Also note the TODO in State.IO#deserialize(): Engine, and so the IO, does not yet support resolving parent states during deserialization.